fix little display digit overlap bug

+++ driver/i2c.h
@@ -31,8 +31,8 @@ void I2C_Stop(void);
 uint8_t I2C_Read(bool bFinal);
 int I2C_Write(uint8_t Data);
 
-int I2C_ReadBuffer(void *pBuffer, uint8_t Size);
-int I2C_WriteBuffer(const void *pBuffer, uint8_t Size);
+int I2C_ReadBuffer(void *pBuffer, unsigned int Size);
+int I2C_WriteBuffer(const void *pBuffer, unsigned int Size);
 
 #endif
